Move integration tests into a separated directory and add script to run only integration tests
