+++ package.json
@@ -55,6 +55,7 @@
     "website": "node docs/generate_website.js",
     "docs": "npm run pdf-guide && npm run website && npm run browser-demo",
     "test": "jest -i",
+    "test:integration": "jest integration/ -i",
     "test:unit": "jest unit/ -i"
   },
   "main": "js/pdfkit.js",
