+++ tests/integration/pdfmake/background.spec.js
@@ -1,5 +1,5 @@
 var {runDocTest} = require('../helpers');
-var {jpeg: bee} = require('../images/bee')
+var {jpeg: bee} = require('../../images/bee')
 
 describe('pdfmake', function () {
   // somehow the generated pdf changes at each run
