+++ tests/integration/pdfmake/basics.spec.js
