+++ tests/integration/pdfmake/columns_simple.spec.js
