+++ tests/integration/pdfmake/qrcode.spec.js
