+++ tests/integration/pdfmake/text_decorations.spec.js
