+++ tests/integration/pdfmake/watermark.spec.js
@@ -1,4 +1,3 @@
-jest.mock('../../lib/security');
 var {runDocTest} = require('../helpers');
 
 describe('pdfmake', function () {
